refactor(html): tree construction "table*"
